frontend/requirements.txt
streamlit==1.10.0
requests==2.28.0
requests-toolbelt==0.9.1
click==8.1.3
Pillow==9.1.1
plotly==5.9.0
scipy==1.8.1
streamlit-folium==0.6.13
joblib==1.1.0
shap==0.41.0
pandas==1.4.2
matplotlib==3.5.2
xgboost==1.6.1